+++ RingSoft.DbLookup.App.Library/Northwind/Model/Product.cs
@@ -28,7 +28,7 @@ namespace RingSoft.DbLookup.App.Library.Northwind.Model
         public string QuantityPerUnit { get; set; }
 
         [Column(TypeName = "money")]
-        public decimal? UnitPrice { get; set; }
+        public double? UnitPrice { get; set; }
 
         public short? UnitsInStock { get; set; }
